requirements.txt
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
decorator==5.1.1
Django==3.2.13
django-ckeditor==6.4.0
django-filter==2.4.0
django-js-asset==2.0.0
django-mongoengine==0.4.6
django-rest-framework-mongoengine==3.4.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djongo==1.3.6
dnspython==2.2.1
drf-spectacular==0.22.0
executing==0.8.3
feedparser==6.0.8
flake8==4.0.1
idna==3.3
importlib-resources==5.6.0
inflection==0.5.1
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.2.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.0
jsonschema==4.4.0
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mongoengine==0.24.1
numpy==1.22.4
openapi-codec==1.3.2
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pluggy==1.0.0
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
pycoingecko==2.2.0
pyflakes==2.4.0
Pygments==2.11.2
PyJWT==2.3.0
pymongo==3.12.3
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.1.1
python-coinmarketcap==0.3
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
sgmllib3k==1.0.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.2.4
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
traitlets==5.1.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
xmltodict==0.12.0
zipp==3.7.0
